feat: Add support for ObsidianYesterday | ObsidianTomorrow for non-workday daily notes (#91)
* feat(config): Add different cadence options to daily notes

This adds a new config field for daily notes: `schedule`.

This is because we currently assume that a daily note is only for a
workweek. On a Sunday, if you used `:ObsidianYesterday`, it returns
Friday's note. In Obsidian, this is not the default behavior.

To not break current behavior, we set `workweek` as the default setting
but allow it to be `calendar` (yesterday is yesterday, tomorrow is
tomorrow, etc).

* feat(util): Add previous and next day helper functions

* feat(client): Use daily note schedule in daily note creation

* refactor: Make daily note schedule boolean

There aren't many conditions where we'd want to have multiple values for
daily note schedules.

To simplify this, let's just make it into a boolean instead.

* feat(util): Add specs for day functions

* docs(daily): Add workweek_only release note

* docs(README): Add workweek_only documentation

* docs(README): clarify ObsidianYesterday vs ObsidianTomorrow

* fix(feedback): workweek_only -> workdays_only