fix(commands): drop complete=file (#109)
* fix(commands): drop complete=file

With `complete=file`, Ex command evaluates substrings quoted by
backticks in its args before the args are passed to their callback
functions via `data`. This should not be an ideal behavior for
`:ObsidianNew` and `ObsidianRename`, and `ObsidianPasteImg`.

Addition to that, the `complete=file` settings do not make sense because
the commands do not expect unique ID of which filename is composed as
arguments, but a title in natural language.

* style(commands): format

* docs(changelog): describe update